修复hc32l19x驱动及soc/hc32系列demo中的部分bug并完善部分文件中的代码规范

+++ soc/hdsc/drivers/include/pca/hw/amhw_hc32_pca.h
@@ -228,7 +228,7 @@ am_static_inline
 void amhw_hc32_pca_int_clear (amhw_hc32_pca_t  *p_hw_pca,
                                 uint8_t             flag)
 {
-    p_hw_pca->iclr &= ~flag;
+    p_hw_pca->iclr &= ~(uint32_t)flag;
 }
 
 /**
